requirements.txt

python-dotenv
pyTelegramBotAPI
schedule
pytz
fpdf